requirements.txt
Pillow
discord.py
jishaku
